feat: WiningLotto 클래스 추가 및 리팩토링
- Lotto 클래스에서 매칭 번호 및 당첨 순위 WinningLotto 클래스로 이동
- 관련 메서드 이동
- Lottery클래스에서 당첨된 로또만 분류하여 리턴하도록 책임 변경

+++ src/test/java/study/lotto/lottery/LotteryResultTest.java
@@ -41,9 +41,9 @@ class LotteryResultTest {
         LotteryResult lotteryResult = lottery.checkLotto();
 
         // Then
-        Integer numberOfLottoRank = lotteryResult.getNumberOfLottoRank(LottoRank.FIRST);
+        Long numberOfLottoRank = lotteryResult.getNumberOfLottoRank(LottoRank.FIRST);
         double totalReturnRatio = lotteryResult.getTotalReturnRatio();
-        assertThat(numberOfLottoRank).isEqualTo(1);
+        assertThat(numberOfLottoRank).isEqualTo(1L);
         assertThat(totalReturnRatio > 1.0).isTrue();
     }
